+++ test/unit/player.js
@@ -211,8 +211,6 @@ test('should set and update the poster value', function(){
   ok(pcEmitted, 'posterchange event was emitted');
   equal(player.poster(), updatedPoster, 'the updated poster is returned');
 
-  equal(player.poster(poster), player, 'the player is returned when setting a poster');
-
   player.dispose();
 });
 
@@ -647,4 +645,12 @@ test('pause is not called if the player is paused and ended is fired', function(
   };
   player.trigger('ended');
   equal(pauses, 0, 'pause was not called when ended fired');
+});
+
+test('should add an audio class if an audio el is used', function() {
+  var audio = document.createElement('audio'),
+      player = PlayerTest.makePlayer({}, audio),
+      audioClass = 'vjs-audio';
+
+  ok(player.el().className.indexOf(audioClass) !== -1, 'added '+ audioClass +' css class');
 });